requirements.txt
backports-abc==0.5
backports.functools-lru-cache==1.4
backports.shutil-get-terminal-size==1.0.0
bleach==2.1.1
certifi==2017.7.27.1
configparser==3.5.0
cycler==0.10.0
decorator==4.1.2
entrypoints==0.2.3
enum34==1.1.6
eyeD3==0.8.2
functools32==3.2.3.post2
hmmlearn==0.2.0
html5lib==1.0b10
hurry.filesize==0.9
ipykernel==4.6.1
ipython==5.5.0
ipython-genutils==0.2.0
Jinja2==2.9.6
jsonschema==2.6.0
jupyter-client==5.1.0
jupyter-core==4.3.0
MarkupSafe==1.0
matplotlib==2.1.0
mistune==0.7.4
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.2.0
numpy==1.13.3
olefile==0.44
pandas==0.20.3
pandocfilters==1.4.2
pathlib==1.0.1
pathlib2==2.3.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.3.0
prompt-toolkit==1.0.15
ptyprocess==0.5.2
pyAudioAnalysis==0.1.3
pydub==0.20.0
Pygments==2.2.0
pyparsing==2.2.0
pypng==0.0.18
python-dateutil==2.6.1
python-magic==0.4.13
pytz==2017.2
pyzmq==16.0.2
scandir==1.6
scikit-learn==0.19.0
scipy==0.19.1
simplegeneric==0.8.1
simplejson==3.11.1
singledispatch==3.4.0.3
six==1.11.0
sklearn==0.0
subprocess32==3.2.7
termcolor==1.1.0
terminado==0.6
testpath==0.3.1
tornado==4.5.2
tqdm==4.19.4
traitlets==4.3.2
wcwidth==0.1.7
webencodings==0.5.1
